Theory: order of configuring the ironic interfaces matters
I think because of routing. The br1 bridge doesn't pass traffic unless
eth1 has an IP, but if that IP is added before br1's, then the jump
host cannot ping the container IPs.

Doing the horrible thing now to verify, and then I can think
harder about the routing and how to do it right.